feat: 루트 문서 추가 버튼 구현
* Sidebar에서 DocumentAddButton 컴포넌트 렌더링

+++ src/components/Sidebar.js
@@ -1,7 +1,11 @@
 import DocumentList from './DocumentList.js';
+import DocumentAddButton from './DocumentAddButton.js';
 
 import { fetchDocuments } from '../utils/api.js';
 import { isNew } from '../utils/helper.js';
+import { NEW, NEW_PARENT, ROUTE_DOCUMENTS } from '../utils/constants.js';
+import { setItem } from '../utils/storage.js';
+import { push } from '../utils/router.js';
 
 export default function Sidebar({ $target }) {
   isNew(new.target);
@@ -15,10 +19,36 @@ export default function Sidebar({ $target }) {
     $target: $sidebar,
     initialState: [],
     onRemove: async (documentId) => {
-      await fetchDocuments(documentId, {
-        method: 'DELETE',
-      });
-      this.render();
+      if (confirm('페이지를 삭제하시겠습니까?')) {
+        await fetchDocuments(documentId, {
+          method: 'DELETE',
+        });
+        this.render();
+      }
+    },
+  });
+
+  new DocumentAddButton({
+    $target: $sidebar,
+    initialState: {
+      position: 'document-list-bottom',
+      text: '페이지 추가',
+    },
+    onClick: () => {
+      setItem(NEW_PARENT, null);
+      push(`${ROUTE_DOCUMENTS}/${NEW}`);
+    },
+  });
+
+  new DocumentAddButton({
+    $target: $sidebar,
+    initialState: {
+      position: 'sidebar-bottom',
+      text: '새 페이지',
+    },
+    onClick: () => {
+      setItem(NEW_PARENT, null);
+      push(`${ROUTE_DOCUMENTS}/${NEW}`);
     },
   });
